Add p50 chart
