+++ static/app/views/starfish/views/spans/spansView.tsx
@@ -9,12 +9,10 @@ import sumBy from 'lodash/sumBy';
 import DatePageFilter from 'sentry/components/datePageFilter';
 import SearchBar from 'sentry/components/searchBar';
 import TagDistributionMeter from 'sentry/components/tagDistributionMeter';
-import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import usePageFilters from 'sentry/utils/usePageFilters';
-import ChartPanel from 'sentry/views/starfish/components/chartPanel';
 import {HOST} from 'sentry/views/starfish/utils/constants';
-import {SpanTotalTimeChart} from 'sentry/views/starfish/views/spans/spanTotalTimeChart';
+import {SpanTimeCharts} from 'sentry/views/starfish/views/spans/spanTimeCharts';
 
 import {CLUSTERS} from './clusters';
 import {getSpanListQuery, getSpansTrendsQuery, getTimeSpentQuery} from './queries';
@@ -184,13 +182,7 @@ export default function SpansView(props: Props) {
         }}
       />
 
-      <ChartsContainer>
-        <ChartsContainerItem>
-          <ChartPanel title={t('Total Time')}>
-            <SpanTotalTimeChart clusters={currentClusters} />
-          </ChartPanel>
-        </ChartsContainerItem>
-      </ChartsContainer>
+      <SpanTimeCharts clusters={currentClusters} />
 
       <SpansTable
         location={props.location}
@@ -210,15 +202,4 @@ const FilterOptionsContainer = styled('div')`
   flex-direction: row;
   gap: ${space(1)};
   margin-bottom: ${space(2)};
-`;
-
-const ChartsContainer = styled('div')`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: ${space(2)};
-`;
-
-const ChartsContainerItem = styled('div')`
-  flex: 1;
 `;
